hecate.nix: do not run tests on macOS
Currently, when running nix-build, tests fail on macOS with the message:

  gpg: can't connect to the agent: File name too long

This is because there is a limit on the length of filename for a local
socket.  When built with Nix, the test suite creates a socket for
gpg-agent in a temporary GNUPGHOME directory somewhere deep within
TMPDIR, exceeding this limit.

Ref:
https://lists.gnupg.org/pipermail/gnupg-users/2017-January/057451.html